Updating the context and making global json
