+++ app/client/src/components/accountInformationPanel/AccountList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AppContext } from '../../AppContext';
+import { AppContext } from '../../context/AppContext';
 import { config } from '../../config';
 import { AccountType } from '../../types/accountTypes';
 import APIDetails from '../APIDetails';
