+++ app/client/src/components/accountInformationPanel/accountCards/AccountCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AppContext } from '../../../AppContext';
+import { AppContext } from '../../../context/AppContext';
 import { AccountType } from '../../../types/accountTypes';
 import { gatherCurrencySymbol, isEmptyObject } from '../../utils';
