+++ app/client/src/components/accountInformationPanel/accountCards/AccountTotal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AppContext } from '../../../AppContext';
+import { AppContext } from '../../../context/AppContext';
 import { config } from '../../../config';
 import { CurrencyType } from '../../../types/accountTypes';
 import APIDetails from '../../APIDetails';
