+++ app/client/src/components/sendPayments/PreviousPaymentsGrid.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function PreviousPaymentsGrid() {
+  return <div className="h-full"> TODO </div>;
+}
+
+export default PreviousPaymentsGrid;
