+++ app/client/src/pages/PaymentsPage.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import SendPaymentForm from '../components/sendPayments/SendPaymentForm';
 import { FormStatus } from '../types/globalPaymentApiTypes';
 import { AccountType } from '../types/accountTypes';
+import PreviousPaymentsGrid from '../components/sendPayments/PreviousPaymentsGrid';
 
 const paymentsAccounts: AccountType[] = [{
   accountId: '000000010900009',
@@ -71,10 +72,14 @@ const paymentsAccounts: AccountType[] = [{
 function PaymentsPage() {
   const [formStatus, setFormStatus] = React.useState<FormStatus>(FormStatus.NEW);
   return (
-    <div className="lg:min-h-screen flex flex-col p-8">
-      <h2 className="text-2xl font-medium mb-4">Send a Payment</h2>
-      <div className="flex flex-wrap flex-grow">
+    <div className="lg:min-h-screen flex p-8 gap-4 w-full flex-row">
+      <div className="w-2/5">
+        <h2 className="text-2xl font-medium mb-4">Send a Payment</h2>
         <SendPaymentForm accountDetails={paymentsAccounts} formStatus={formStatus} setFormStatus={setFormStatus} />
+      </div>
+      <div className="flex-grow gap-4">
+        <h2 className="text-2xl font-medium mb-4">Previous payments</h2>
+        <PreviousPaymentsGrid />
       </div>
     </div>
   );
